+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Since last release
 * Added TransportUnits (#1750, #1772, #1784)
 * Added CI support for Ubuntu 24.04 (#1770)
 * Added ``doxygen`` to list of dependencies installed in the Dockerfile (#1782)
+* Added macros in ``cmake/CyclusBuildSetup.cmake`` for common CMake boilerplate (#1793)
 
 **Changed:**